Refactor and creates Flow/Utils library
This pulls in common code needed at dispatch and codegen time.
This is a medium term solution to gather common methods before
we use op interfaces.
