+++ iree/compiler/Codegen/LLVMGPU/CMakeLists.txt
@@ -61,6 +61,7 @@ iree_cc_library(
     iree::compiler::Codegen::PassHeaders
     iree::compiler::Codegen::Transforms
     iree::compiler::Codegen::Utils
+    iree::compiler::Dialect::Flow::Utils
     iree::compiler::Dialect::HAL::IR
     iree::compiler::Dialect::IREE::IR
     iree::compiler::Dialect::LinalgExt::IR
