+++ iree/compiler/Dialect/Flow/Transforms/BUILD
@@ -66,6 +66,7 @@ cc_library(
     deps = [
         ":PassesIncGen",
         "//iree/compiler/Dialect/Flow/IR",
+        "//iree/compiler/Dialect/Flow/Utils",
         "//iree/compiler/Dialect/HAL/IR",
         "//iree/compiler/Dialect/IREE/IR",
         "//iree/compiler/Dialect/LinalgExt/IR",
