+++ iree/compiler/Dialect/Flow/Utils/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# iree/compiler/Codegen/Utils/BUILD                                            #
+# iree/compiler/Dialect/Flow/Utils/BUILD                                       #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -14,20 +14,13 @@ iree_cc_library(
   NAME
     Utils
   HDRS
-    "MarkerUtils.h"
-    "Utils.h"
+    "DispatchUtils.h"
   SRCS
-    "MarkerUtils.cpp"
-    "Utils.cpp"
+    "DispatchUtils.cpp"
   DEPS
-    LLVMSupport
     MLIRIR
     MLIRLinalg
     MLIRLinalgTransforms
-    MLIRSupport
-    iree::compiler::Dialect::Flow::IR
-    iree::compiler::Dialect::Flow::Utils
-    iree::compiler::Dialect::HAL::IR
     iree::compiler::Dialect::LinalgExt::IR
   PUBLIC
 )
